Cleanup

+++ src/autosaveworld/zlibs/com/fasterxml/jackson/core/Base64Variant.java
@@ -14,7 +14,7 @@ import autosaveworld.zlibs.com.fasterxml.jackson.core.util.ByteArrayBuilder;
  * basic version (so-called "MIME Base64"), other variants exists, see <a
  * href="http://en.wikipedia.org/wiki/Base64">Base64 Wikipedia entry</a> for
  * details.
- * 
+ *
  * @author Tatu Saloranta
  */
 public final class Base64Variant implements java.io.Serializable {
@@ -130,7 +130,7 @@ public final class Base64Variant implements java.io.Serializable {
 
 		// Plus if we use padding, add that in too
 		if (usesPadding) {
-			_asciiToBase64[(int) paddingChar] = BASE64_VALUE_PADDING;
+			_asciiToBase64[paddingChar] = BASE64_VALUE_PADDING;
 		}
 	}
 
@@ -194,7 +194,7 @@ public final class Base64Variant implements java.io.Serializable {
 	}
 
 	public boolean usesPaddingChar(int ch) {
-		return ch == (int) _paddingChar;
+		return ch == _paddingChar;
 	}
 
 	public char getPaddingChar() {
@@ -218,7 +218,7 @@ public final class Base64Variant implements java.io.Serializable {
 	 * @return 6-bit decoded value, if valid character;
 	 */
 	public int decodeBase64Char(char c) {
-		int ch = (int) c;
+		int ch = c;
 		return (ch <= 127) ? _asciiToBase64[ch] : BASE64_VALUE_INVALID;
 	}
 
@@ -227,7 +227,7 @@ public final class Base64Variant implements java.io.Serializable {
 	}
 
 	public int decodeBase64Byte(byte b) {
-		int ch = (int) b;
+		int ch = b;
 		return (ch <= 127) ? _asciiToBase64[ch] : BASE64_VALUE_INVALID;
 	}
 
@@ -267,7 +267,7 @@ public final class Base64Variant implements java.io.Serializable {
 	 * Method that outputs partial chunk (which only encodes one or two bytes of
 	 * data). Data given is still aligned same as if it as full data; that is,
 	 * missing data is at the "right end" (LSB) of int.
-	 * 
+	 *
 	 * @param outputBytes
 	 *            Number of encoded bytes included (either 1 or 2)
 	 */
@@ -322,7 +322,7 @@ public final class Base64Variant implements java.io.Serializable {
 	 * Method that outputs partial chunk (which only encodes one or two bytes of
 	 * data). Data given is still aligned same as if it as full data; that is,
 	 * missing data is at the "right end" (LSB) of int.
-	 * 
+	 *
 	 * @param outputBytes
 	 *            Number of encoded bytes included (either 1 or 2)
 	 */
@@ -353,7 +353,7 @@ public final class Base64Variant implements java.io.Serializable {
 	 * Convenience method for converting given byte array as base64 encoded
 	 * String using this variant's settings. Resulting value is "raw", that is,
 	 * not enclosed in double-quotes.
-	 * 
+	 *
 	 * @param input
 	 *            Byte array to encode
 	 */
@@ -365,7 +365,7 @@ public final class Base64Variant implements java.io.Serializable {
 	 * Convenience method for converting given byte array as base64 encoded
 	 * String using this variant's settings, optionally enclosed in
 	 * double-quotes.
-	 * 
+	 *
 	 * @param input
 	 *            Byte array to encode
 	 * @param addQuotes
@@ -391,9 +391,9 @@ public final class Base64Variant implements java.io.Serializable {
 
 		while (inputPtr <= safeInputEnd) {
 			// First, mash 3 bytes into lsb of 32-bit int
-			int b24 = ((int) input[inputPtr++]) << 8;
-			b24 |= ((int) input[inputPtr++]) & 0xFF;
-			b24 = (b24 << 8) | (((int) input[inputPtr++]) & 0xFF);
+			int b24 = (input[inputPtr++]) << 8;
+			b24 |= (input[inputPtr++]) & 0xFF;
+			b24 = (b24 << 8) | ((input[inputPtr++]) & 0xFF);
 			encodeBase64Chunk(sb, b24);
 			if (--chunksBeforeLF <= 0) {
 				// note: must quote in JSON value, so not really useful...
@@ -406,9 +406,9 @@ public final class Base64Variant implements java.io.Serializable {
 		// And then we may have 1 or 2 leftover bytes to encode
 		int inputLeft = inputEnd - inputPtr; // 0, 1 or 2
 		if (inputLeft > 0) { // yes, but do we have room for output?
-			int b24 = ((int) input[inputPtr++]) << 16;
+			int b24 = (input[inputPtr++]) << 16;
 			if (inputLeft == 2) {
-				b24 |= (((int) input[inputPtr++]) & 0xFF) << 8;
+				b24 |= ((input[inputPtr++]) & 0xFF) << 8;
 			}
 			encodeBase64Partial(sb, b24, inputLeft);
 		}
@@ -422,11 +422,11 @@ public final class Base64Variant implements java.io.Serializable {
 	/**
 	 * Convenience method for decoding contents of a Base64-encoded String,
 	 * using this variant's settings.
-	 * 
+	 *
 	 * @param input
-	 * 
+	 *
 	 * @since 2.2.3
-	 * 
+	 *
 	 * @throws IllegalArgumentException
 	 *             if input is not valid base64 encoded data
 	 */
@@ -444,9 +444,9 @@ public final class Base64Variant implements java.io.Serializable {
 	 * NOTE: builder will NOT be reset before decoding (nor cleared afterwards);
 	 * assumption is that caller will ensure it is given in proper state, and
 	 * used as appropriate afterwards.
-	 * 
+	 *
 	 * @since 2.2.3
-	 * 
+	 *
 	 * @throws IllegalArgumentException
 	 *             if input is not valid base64 encoded data
 	 */
