+++ src/autosaveworld/zlibs/com/fasterxml/jackson/core/json/UTF8JsonGenerator.java
@@ -1,11 +1,20 @@
 package autosaveworld.zlibs.com.fasterxml.jackson.core.json;
 
-import java.io.*;
+import java.io.IOException;
+import java.io.InputStream;
+import java.io.OutputStream;
 import java.math.BigDecimal;
 import java.math.BigInteger;
 
-import autosaveworld.zlibs.com.fasterxml.jackson.core.*;
-import autosaveworld.zlibs.com.fasterxml.jackson.core.io.*;
+import autosaveworld.zlibs.com.fasterxml.jackson.core.Base64Variant;
+import autosaveworld.zlibs.com.fasterxml.jackson.core.JsonGenerationException;
+import autosaveworld.zlibs.com.fasterxml.jackson.core.JsonStreamContext;
+import autosaveworld.zlibs.com.fasterxml.jackson.core.ObjectCodec;
+import autosaveworld.zlibs.com.fasterxml.jackson.core.SerializableString;
+import autosaveworld.zlibs.com.fasterxml.jackson.core.io.CharTypes;
+import autosaveworld.zlibs.com.fasterxml.jackson.core.io.CharacterEscapes;
+import autosaveworld.zlibs.com.fasterxml.jackson.core.io.IOContext;
+import autosaveworld.zlibs.com.fasterxml.jackson.core.io.NumberOutput;
 
 public class UTF8JsonGenerator extends JsonGeneratorImpl {
 	private final static byte BYTE_u = (byte) 'u';
@@ -609,7 +618,7 @@ public class UTF8JsonGenerator extends JsonGeneratorImpl {
 		// checks
 		main_loop: while (offset < len) {
 			inner_loop: while (true) {
-				int ch = (int) cbuf[offset];
+				int ch = cbuf[offset];
 				if (ch > 0x7F) {
 					break inner_loop;
 				}
@@ -655,7 +664,7 @@ public class UTF8JsonGenerator extends JsonGeneratorImpl {
 
 		main_loop: while (offset < len) {
 			inner_loop: while (true) {
-				int ch = (int) cbuf[offset];
+				int ch = cbuf[offset];
 				if (ch >= 0x80) {
 					break inner_loop;
 				}
@@ -1518,9 +1527,9 @@ public class UTF8JsonGenerator extends JsonGeneratorImpl {
 				_flushBuffer();
 			}
 			// First, mash 3 bytes into lsb of 32-bit int
-			int b24 = ((int) input[inputPtr++]) << 8;
-			b24 |= ((int) input[inputPtr++]) & 0xFF;
-			b24 = (b24 << 8) | (((int) input[inputPtr++]) & 0xFF);
+			int b24 = (input[inputPtr++]) << 8;
+			b24 |= (input[inputPtr++]) & 0xFF;
+			b24 = (b24 << 8) | ((input[inputPtr++]) & 0xFF);
 			_outputTail = b64variant.encodeBase64Chunk(b24, _outputBuffer,
 					_outputTail);
 			if (--chunksBeforeLF <= 0) {
@@ -1538,9 +1547,9 @@ public class UTF8JsonGenerator extends JsonGeneratorImpl {
 												// but...
 				_flushBuffer();
 			}
-			int b24 = ((int) input[inputPtr++]) << 16;
+			int b24 = (input[inputPtr++]) << 16;
 			if (inputLeft == 2) {
-				b24 |= (((int) input[inputPtr++]) & 0xFF) << 8;
+				b24 |= ((input[inputPtr++]) & 0xFF) << 8;
 			}
 			_outputTail = b64variant.encodeBase64Partial(b24, inputLeft,
 					_outputBuffer, _outputTail);
@@ -1573,9 +1582,9 @@ public class UTF8JsonGenerator extends JsonGeneratorImpl {
 			if (_outputTail > safeOutputEnd) { // need to flush
 				_flushBuffer();
 			}
-			int b24 = ((int) readBuffer[inputPtr++]) << 8;
-			b24 |= ((int) readBuffer[inputPtr++]) & 0xFF;
-			b24 = (b24 << 8) | (((int) readBuffer[inputPtr++]) & 0xFF);
+			int b24 = (readBuffer[inputPtr++]) << 8;
+			b24 |= (readBuffer[inputPtr++]) & 0xFF;
+			b24 = (b24 << 8) | ((readBuffer[inputPtr++]) & 0xFF);
 			bytesLeft -= 3;
 			_outputTail = b64variant.encodeBase64Chunk(b24, _outputBuffer,
 					_outputTail);
@@ -1596,10 +1605,10 @@ public class UTF8JsonGenerator extends JsonGeneratorImpl {
 													// but...
 					_flushBuffer();
 				}
-				int b24 = ((int) readBuffer[inputPtr++]) << 16;
+				int b24 = (readBuffer[inputPtr++]) << 16;
 				int amount;
 				if (inputPtr < inputEnd) {
-					b24 |= (((int) readBuffer[inputPtr]) & 0xFF) << 8;
+					b24 |= ((readBuffer[inputPtr]) & 0xFF) << 8;
 					amount = 2;
 				} else {
 					amount = 1;
@@ -1641,9 +1650,9 @@ public class UTF8JsonGenerator extends JsonGeneratorImpl {
 				_flushBuffer();
 			}
 			// First, mash 3 bytes into lsb of 32-bit int
-			int b24 = ((int) readBuffer[inputPtr++]) << 8;
-			b24 |= ((int) readBuffer[inputPtr++]) & 0xFF;
-			b24 = (b24 << 8) | (((int) readBuffer[inputPtr++]) & 0xFF);
+			int b24 = (readBuffer[inputPtr++]) << 8;
+			b24 |= (readBuffer[inputPtr++]) & 0xFF;
+			b24 = (b24 << 8) | ((readBuffer[inputPtr++]) & 0xFF);
 			bytesDone += 3;
 			_outputTail = b64variant.encodeBase64Chunk(b24, _outputBuffer,
 					_outputTail);
@@ -1660,10 +1669,10 @@ public class UTF8JsonGenerator extends JsonGeneratorImpl {
 												// but...
 				_flushBuffer();
 			}
-			int b24 = ((int) readBuffer[inputPtr++]) << 16;
+			int b24 = (readBuffer[inputPtr++]) << 16;
 			int amount = 1;
 			if (inputPtr < inputEnd) {
-				b24 |= (((int) readBuffer[inputPtr]) & 0xFF) << 8;
+				b24 |= ((readBuffer[inputPtr]) & 0xFF) << 8;
 				amount = 2;
 			}
 			bytesDone += amount;
@@ -1744,13 +1753,13 @@ public class UTF8JsonGenerator extends JsonGeneratorImpl {
 	}
 
 	/**
-	 * 
+	 *
 	 * @param ch
 	 * @param outputPtr
 	 *            Position within output buffer to append multi-byte in
-	 * 
+	 *
 	 * @return New output position after appending
-	 * 
+	 *
 	 * @throws IOException
 	 */
 	private final int _outputMultiByteChar(int ch, int outputPtr)
@@ -1796,7 +1805,7 @@ public class UTF8JsonGenerator extends JsonGeneratorImpl {
 
 	/**
 	 * Method called to write a generic Unicode escape for given character.
-	 * 
+	 *
 	 * @param charToEscape
 	 *            Character to escape using escape sequence (\\uXXXX)
 	 */
