+++ src/autosaveworld/zlibs/com/fasterxml/jackson/core/json/WriterBasedJsonGenerator.java
@@ -1,11 +1,21 @@
 package autosaveworld.zlibs.com.fasterxml.jackson.core.json;
 
-import java.io.*;
+import java.io.IOException;
+import java.io.InputStream;
+import java.io.Writer;
 import java.math.BigDecimal;
 import java.math.BigInteger;
 
-import autosaveworld.zlibs.com.fasterxml.jackson.core.*;
-import autosaveworld.zlibs.com.fasterxml.jackson.core.io.*;
+import autosaveworld.zlibs.com.fasterxml.jackson.core.Base64Variant;
+import autosaveworld.zlibs.com.fasterxml.jackson.core.JsonGenerationException;
+import autosaveworld.zlibs.com.fasterxml.jackson.core.JsonGenerator;
+import autosaveworld.zlibs.com.fasterxml.jackson.core.JsonStreamContext;
+import autosaveworld.zlibs.com.fasterxml.jackson.core.ObjectCodec;
+import autosaveworld.zlibs.com.fasterxml.jackson.core.SerializableString;
+import autosaveworld.zlibs.com.fasterxml.jackson.core.io.CharTypes;
+import autosaveworld.zlibs.com.fasterxml.jackson.core.io.CharacterEscapes;
+import autosaveworld.zlibs.com.fasterxml.jackson.core.io.IOContext;
+import autosaveworld.zlibs.com.fasterxml.jackson.core.io.NumberOutput;
 
 /**
  * {@link JsonGenerator} that outputs JSON content using a
@@ -1395,9 +1405,9 @@ public final class WriterBasedJsonGenerator extends JsonGeneratorImpl {
 				_flushBuffer();
 			}
 			// First, mash 3 bytes into lsb of 32-bit int
-			int b24 = ((int) input[inputPtr++]) << 8;
-			b24 |= ((int) input[inputPtr++]) & 0xFF;
-			b24 = (b24 << 8) | (((int) input[inputPtr++]) & 0xFF);
+			int b24 = (input[inputPtr++]) << 8;
+			b24 |= (input[inputPtr++]) & 0xFF;
+			b24 = (b24 << 8) | ((input[inputPtr++]) & 0xFF);
 			_outputTail = b64variant.encodeBase64Chunk(b24, _outputBuffer,
 					_outputTail);
 			if (--chunksBeforeLF <= 0) {
@@ -1415,9 +1425,9 @@ public final class WriterBasedJsonGenerator extends JsonGeneratorImpl {
 												// but...
 				_flushBuffer();
 			}
-			int b24 = ((int) input[inputPtr++]) << 16;
+			int b24 = (input[inputPtr++]) << 16;
 			if (inputLeft == 2) {
-				b24 |= (((int) input[inputPtr++]) & 0xFF) << 8;
+				b24 |= ((input[inputPtr++]) & 0xFF) << 8;
 			}
 			_outputTail = b64variant.encodeBase64Partial(b24, inputLeft,
 					_outputBuffer, _outputTail);
@@ -1450,9 +1460,9 @@ public final class WriterBasedJsonGenerator extends JsonGeneratorImpl {
 			if (_outputTail > safeOutputEnd) { // need to flush
 				_flushBuffer();
 			}
-			int b24 = ((int) readBuffer[inputPtr++]) << 8;
-			b24 |= ((int) readBuffer[inputPtr++]) & 0xFF;
-			b24 = (b24 << 8) | (((int) readBuffer[inputPtr++]) & 0xFF);
+			int b24 = (readBuffer[inputPtr++]) << 8;
+			b24 |= (readBuffer[inputPtr++]) & 0xFF;
+			b24 = (b24 << 8) | ((readBuffer[inputPtr++]) & 0xFF);
 			bytesLeft -= 3;
 			_outputTail = b64variant.encodeBase64Chunk(b24, _outputBuffer,
 					_outputTail);
@@ -1473,10 +1483,10 @@ public final class WriterBasedJsonGenerator extends JsonGeneratorImpl {
 													// but...
 					_flushBuffer();
 				}
-				int b24 = ((int) readBuffer[inputPtr++]) << 16;
+				int b24 = (readBuffer[inputPtr++]) << 16;
 				int amount;
 				if (inputPtr < inputEnd) {
-					b24 |= (((int) readBuffer[inputPtr]) & 0xFF) << 8;
+					b24 |= ((readBuffer[inputPtr]) & 0xFF) << 8;
 					amount = 2;
 				} else {
 					amount = 1;
@@ -1517,9 +1527,9 @@ public final class WriterBasedJsonGenerator extends JsonGeneratorImpl {
 				_flushBuffer();
 			}
 			// First, mash 3 bytes into lsb of 32-bit int
-			int b24 = ((int) readBuffer[inputPtr++]) << 8;
-			b24 |= ((int) readBuffer[inputPtr++]) & 0xFF;
-			b24 = (b24 << 8) | (((int) readBuffer[inputPtr++]) & 0xFF);
+			int b24 = (readBuffer[inputPtr++]) << 8;
+			b24 |= (readBuffer[inputPtr++]) & 0xFF;
+			b24 = (b24 << 8) | ((readBuffer[inputPtr++]) & 0xFF);
 			bytesDone += 3;
 			_outputTail = b64variant.encodeBase64Chunk(b24, _outputBuffer,
 					_outputTail);
@@ -1536,10 +1546,10 @@ public final class WriterBasedJsonGenerator extends JsonGeneratorImpl {
 												// but...
 				_flushBuffer();
 			}
-			int b24 = ((int) readBuffer[inputPtr++]) << 16;
+			int b24 = (readBuffer[inputPtr++]) << 16;
 			int amount = 1;
 			if (inputPtr < inputEnd) {
-				b24 |= (((int) readBuffer[inputPtr]) & 0xFF) << 8;
+				b24 |= ((readBuffer[inputPtr]) & 0xFF) << 8;
 				amount = 2;
 			}
 			bytesDone += amount;
@@ -1691,7 +1701,7 @@ public final class WriterBasedJsonGenerator extends JsonGeneratorImpl {
 	/**
 	 * Method called to try to either prepend character escape at front of given
 	 * buffer; or if not possible, to write it out directly.
-	 * 
+	 *
 	 * @return Pointer to start of prepended entity (if prepended); or 'ptr' if
 	 *         not.
 	 */
